store_analytics/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
pytest
pytest-asyncio
black
isort
mypy
flake8
